envite/requirements.txt
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
dj-rest-auth==7.0.1
Django==4.2.21
django-cors-headers==4.4.0
django-hashid-field==3.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
easy-thumbnails==2.9
gunicorn==23.0.0
hashids==1.3.1
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==25.0
pillow==10.4.0
pkgutil_resolve_name==1.3.10
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
zipp==3.20.2
